FX/Phase1/requirements.txt
requests
feedparser
readability-lxml
beautifulsoup4
langdetect
tldextract
python-dotenv
lxml
